triggers/gitbot/requirements.txt
discord>=1.7.2
ghapi>=0.1.17
pyyaml==5.4.1
python-jenkins==1.7.0